move store loader to core as well
